% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat2df.R
\name{mat2df}
\alias{mat2df}
\title{Matrix to data frame}
\usage{
mat2df(network)
}
\arguments{
\item{network}{a matrix}
}
\value{
a data frame
}
\description{
Matrix to data frame
}
\examples{
net <- matrix(c(5,3,5,3,1,1,
                3,3,3,3,0,0,
                3,0,0,0,1,0,
                0,0,3,1,0,0), byrow = TRUE, nrow = 4, ncol = 6,
               dimnames = list(paste('P',1:4,sep=''),paste('A',1:6,sep='')))
mat2df(net)

}
