% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_monolayer.R
\name{plot_monolayer}
\alias{plot_monolayer}
\title{Plot bipartite networks}
\usage{
plot_monolayer(
  network,
  color_low = "#00AFBB",
  color_high = "#E7B800",
  shape_low = 19,
  shape_high = 19,
  color_link = "grey80",
  link_type = 1,
  binary = FALSE
)
}
\arguments{
\item{network}{a bipartite network matrix}

\item{color_low}{color of species in the lower trophic level (matrix row)}

\item{color_high}{color of species in the higher trophic level (matrix col)}

\item{shape_low}{node shape of species in the lower trophic level}

\item{shape_high}{node shape of of species in the higher trophic level}

\item{color_link}{color of links}

\item{link_type}{line type of links}

\item{binary}{TRUE or FALSE: return a binary or weighted network}
}
\value{
ggplot2 object
}
\description{
Plot bipartite networks
}
\examples{
pollination <- matrix(c(5,3,5,3,1,1,
                        3,3,3,3,0,0,
                        3,0,0,0,1,0,
                        0,0,3,1,0,0), byrow = TRUE, nrow = 4, ncol = 6,
dimnames = list(paste('P',1:4,sep=''), paste('A',1:6,sep='')))

library(ggplot2)
plot_monolayer(pollination[names(sort(rowSums(pollination), decreasing = TRUE)),
names(sort(colSums(pollination), decreasing = TRUE))],
shape_low = 19, shape_high = 19, color_link = "#FC4E07",
link_type = 1, binary = FALSE) + coord_flip()

}
