% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spatial_multilayer.R
\name{plot_spatial_multilayer}
\alias{plot_spatial_multilayer}
\title{Plot out spatial multilayer networks}
\usage{
plot_spatial_multilayer(
  networks,
  interlayer,
  sequence = NULL,
  symmetry_inter = FALSE,
  binary = FALSE,
  y_adjust = 1,
  y_adjust_h = 1.5,
  gap = 0.6,
  label = FALSE,
  extend = NULL,
  inter_lty = 2,
  inter_lcol = "black",
  intra_lty = 1,
  intra_lcol = "black",
  inter_lsize = 1,
  intra_lsize = 1,
  node_size = 1,
  label_size = 1.5,
  alpha = 0.2,
  ...
)
}
\arguments{
\item{networks}{a list object of interaction network matrices}

\item{interlayer}{a data frame with `layer_from`,`node_from`,`layer_to`, `node_to`, and `weight` columns.}

\item{sequence}{a list}

\item{symmetry_inter}{TRUE or FALSE}

\item{binary}{plot a binary network or a weighted network}

\item{y_adjust}{control the shape of polygon}

\item{y_adjust_h}{control the shape of polygon}

\item{gap}{overlap between layers}

\item{label}{node label}

\item{extend}{extend information about nodes}

\item{inter_lty}{line type of inter-links}

\item{inter_lcol}{line color of inter-links}

\item{intra_lty}{line type of intra-links}

\item{intra_lcol}{line color of intra-links}

\item{inter_lsize}{line size of inter-links}

\item{intra_lsize}{line size of intra-links}

\item{node_size}{node size}

\item{label_size}{size of labels}

\item{alpha}{transparency of layers}

\item{...}{to ggplot2}
}
\value{
ggplot object
}
\description{
Plot out spatial multilayer networks
}
\examples{
site_1 <- matrix(c(5,3,5,3,1,1,
                   3,3,3,3,0,0,
                   3,0,0,0,1,0,
                   0,0,3,1,0,0), byrow = TRUE, nrow = 4, ncol = 6,
                   dimnames = list(paste('P',1:4,sep=''),paste('A',1:6,sep='')))

site_2 <- matrix(c(7,0,5,3,0,1,
                   5,2,3,2,0,1,
                   1,0,0,0,1,0,
                   0,2,2,1,0,0), byrow = TRUE, nrow = 4, ncol = 6,
                 dimnames = list(paste('P',1:4,sep=''),paste('A',1:6,sep='')))

site_3 <- matrix(c(5,1,1,1,0,1,
                   3,0,1,1,0,1,
                   2,0,0,0,0,0,
                   1,0,0,1,0,0), byrow = TRUE, nrow = 4, ncol = 6,
                 dimnames = list(paste('P', 1:4,sep=''), paste('A', 1:6,sep='')))



spatial_nets <- list(site_1 = site_1, site_2 = site_2, site_3 = site_3)

interlayer_space <- dplyr::tibble(layer_from = c(1, 2, 1, 2, 1, 1),
                           node_from = c('P1', 'P1', 'P1', 'P4', 'A3', 'A3'),
                           layer_to =  c(2, 3, 3, 3, 2, 3),
                           node_to =  c('P1',  'P1', 'P1', 'P4', 'A3', 'A3'),
                           weight = c(5, 4, 1, 2, 1, 2))

interlayer_space_2 <- dplyr::tibble(layer_from = c(1, 2, 3, 2),
                             node_from = c('P1', 'P1', 'P1', 'P4'),
                             layer_to =  c(2, 3, 1, 3),
                             node_to =  c('P1',  'P1', 'P1', 'P4'),
                             weight = c(5, 4, 1, 2))
library(ggplot2)
plot_spatial_multilayer(networks = spatial_nets, interlayer = interlayer_space,
                        symmetry_inter = TRUE, binary = TRUE, y_adjust = 1,
                        y_adjust_h = 1, gap = 0.6, label = FALSE,
                        extend = NULL, inter_lty = 1, inter_lcol = "#00AFBB",
                        intra_lty = 1, intra_lcol = "#E7B800", inter_lsize = 0.5,
                        intra_lsize = 1, node_size = 4, label_size = 1.5)+
  scale_fill_manual(values = c("#00AFBB", "#E7B800", "#FC4E07"))


plot_spatial_multilayer(networks = spatial_nets, interlayer = interlayer_space_2,
                        symmetry_inter = FALSE, binary = TRUE, y_adjust = 1,
                        y_adjust_h = 1, gap = 0.6, label = FALSE,
                        extend = NULL, inter_lty = 1, inter_lcol = "#00AFBB",
                        intra_lty = 1, intra_lcol = "#E7B800", inter_lsize = 0.5,
                        intra_lsize = 1, node_size = 4, label_size = 1.5)+
  scale_fill_manual(values = c("#00AFBB", "#E7B800", "#FC4E07"))


}
