% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_temporal_multilayer.R
\name{plot_temporal_multilayer}
\alias{plot_temporal_multilayer}
\title{Plot out a temporal multilayer network}
\usage{
plot_temporal_multilayer(
  networks,
  interlayer,
  sequence = NULL,
  binary = FALSE,
  y_adjust = 1,
  y_adjust_h = 1.5,
  gap = 0.6,
  label = FALSE,
  extend = NULL,
  inter_lty = 2,
  inter_lcol = "black",
  intra_lty = 1,
  intra_lcol = "black",
  inter_lsize = 1,
  intra_lsize = 1,
  node_size = 1,
  label_size = 1.5,
  alpha = 0.2,
  ...
)
}
\arguments{
\item{networks}{a list object of interaction network matrices}

\item{interlayer}{a data frame with `layer_from`,`node_from`,`layer_to`, `node_to`, and `weight` columns.}

\item{sequence}{a list}

\item{binary}{plot a binary network or a weighted network}

\item{y_adjust}{control the shape of polygon}

\item{y_adjust_h}{control the shape of polygon}

\item{gap}{overlap between layers}

\item{label}{node label}

\item{extend}{extend information about nodes}

\item{inter_lty}{line type of inter-links}

\item{inter_lcol}{line color of inter-links}

\item{intra_lty}{line type of intra-links}

\item{intra_lcol}{line color of intra-links}

\item{inter_lsize}{line size of inter-links}

\item{intra_lsize}{line size of intra-links}

\item{node_size}{node size}

\item{label_size}{size of labels}

\item{alpha}{transparency of layers}

\item{...}{to ggplot2}
}
\value{
ggplot object
}
\description{
Plot out a temporal multilayer network
}
\examples{
time_1 <- matrix(c(5,3,5,3,1,1,
                   3,3,3,3,0,0,
                   3,0,0,0,1,0,
                   0,0,3,1,0,0), byrow = TRUE, nrow = 4, ncol = 6,
                 dimnames = list(paste('P',1:4,sep=''),paste('A',1:6,sep='')))

time_2 <- matrix(c(7,0,5,3,0,1,
                  5,2,3,2,0,1,
                  1,0,0,0,1,0,
                  0,2,2,1,0,0), byrow = TRUE, nrow = 4, ncol = 6,
                 dimnames = list(paste('P',1:4,sep=''),paste('A',1:6,sep='')))

time_3 <- matrix(c(5,1,1,1,0,1,
                  3,0,1,1,0,1,
                   2,0,0,0,0,0,
                  1,0,0,1,0,0), byrow = TRUE, nrow = 4, ncol = 6,
                dimnames = list(paste('P',1:4,sep=''),paste('A',1:6,sep='')))



time_nets <- list(time_1 = time_1, time_2 = time_2, time_3 = time_3)

interlayer <- dplyr::tibble(layer_from = c(1, 2, 2, 2),
                    node_from = c('P1', 'P1', 'P2', 'P4'),
                     layer_to =  c(2, 3, 3, 3),
                    node_to =  c('P1',  'P1', 'P2', 'P4'),
                    weight = c(5, 4, 1, 2))

library(ggplot2)
plot_temporal_multilayer(networks = time_nets, interlayer = interlayer,
                         sequence = list(c('P1', 'P3', 'P2', 'P4'), c('A3','A5' ,'A1', 'A2', 'A4', 'A6')),
                        binary = TRUE,node_size = 4, intra_lcol = "#00AFBB", label = TRUE,
                        inter_lcol = "#E7B800", inter_lty = 5, y_adjust = 2) +
  scale_color_manual(values = c("#00AFBB", "#E7B800")) +
  scale_fill_manual(values = c("#00AFBB", "#E7B800", "#FC4E07"))


}
