% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tripartite.R
\name{plot_tripartite}
\alias{plot_tripartite}
\title{plot tripartite networks}
\usage{
plot_tripartite(
  network1,
  network2,
  color_shared = "#00AFBB",
  color_1 = "#E7B800",
  color_2 = "#FC4E07",
  color_link_1 = "grey80",
  color_link_2 = "grey80",
  shape_shared = 19,
  shape_1 = 19,
  shape_2 = 19,
  link_lty_1 = 1,
  link_lty_2 = 1,
  link_size_1 = 1,
  link_size_2 = 1,
  binary = FALSE,
  label = FALSE,
  label_size = 2,
  node_size = 4
)
}
\arguments{
\item{network1}{a network matrix}

\item{network2}{another network matrix}

\item{color_shared}{color of shared species between the two networks}

\item{color_1}{color of nodes (none shared) of network 1}

\item{color_2}{color of nodes (none shared) of network 2}

\item{color_link_1}{color of intra-links of layer 1}

\item{color_link_2}{color of intra-links of layer 2}

\item{shape_shared}{point shape of shared species between the two networks}

\item{shape_1}{point shape of nodes (none shared) of network 1}

\item{shape_2}{point shape of nodes (none shared) of network 2}

\item{link_lty_1}{line type of intra-links of network 1}

\item{link_lty_2}{line type of intra-links of network 2}

\item{link_size_1}{line size of intra-links of network 1}

\item{link_size_2}{line size of intra-links of network 2}

\item{binary}{TRUE or FALSE: return a binary or weighted network}

\item{label}{TRUE or FALSE: nodes with label or not, default is FALSE}

\item{label_size}{label size}

\item{node_size}{node size}
}
\value{
ggplot2 object
}
\description{
plot tripartite networks
}
\examples{
pollination <- matrix(c(5,3,5,3,1,1,
                        3,3,3,3,0,0,
                        3,0,0,0,1,0,
                        0,0,3,1,0,0), byrow = TRUE, nrow = 4, ncol = 6,
dimnames = list(paste('P',1:4,sep=''), paste('A',1:6,sep='')))

herbivory <- matrix(c(7,0,5,3,0,1,
                      5,2,3,2,0,1,
                      1,0,0,0,1,0,
                      0,2,2,1,0,0), byrow = TRUE, nrow = 4, ncol = 6,
                    dimnames = list(paste('P',1:4,sep=''),paste('H',1:6,sep='')))
library(ggplot2)
plot_tripartite(pollination, herbivory, binary = FALSE, label = FALSE) +
coord_flip(expand = TRUE)
}
